requirements.txt
openai
replicate
pika
psycopg2-binary
PyYAML
pytest
requests
